% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_glmnet.R
\name{animate_glmnet}
\alias{animate_glmnet}
\title{Animate glmnet}
\usage{
animate_glmnet(cv.glmnet, plot.cvm = TRUE, plot.cv.folds = TRUE,
  total.time = 10, new.save = TRUE, save.html = TRUE, debug = FALSE,
  debug.n = 100, ...)
}
\arguments{
\item{cv.glmnet}{An object of class 'cv.glmnet'}

\item{plot.cvm}{Should cross-validation error be plotted? Defaults to TRUE.}

\item{plot.cv.folds}{Should cross-validation folds be plotted? Defaults to TRUE.}

\item{total.time}{Desired time of animation in seconds. Defaults to 10.}

\item{new.save}{Should this animation be saved as a new object rather than overwrite the preceeding animation? Defaults to TRUE.}

\item{save.html}{Save as HTML? Defaults to TRUE. If FALSE, saves GIF.}

\item{debug}{Only plot subset of lambda values? Defaults to FALSE.}

\item{debug.n}{If plotting subset of lambda values, sets number of values to plot. Defaults to 100.}

\item{...}{Options passed to saveHTML or saveGIF functions. See ?animate::saveHTML and ?animate::saveGIF}
}
\description{
This function visualizes the a cv.glmnet machine learning process.
}
\examples{
# See also: ?cv.glmnet: 
set.seed(1010)
n=1000;p=100
nzc=trunc(p/10)
x=matrix(rnorm(n*p),n,p)
beta=rnorm(nzc)
fx= x[,seq(nzc)] \%*\% beta
eps=rnorm(n)*5
y=drop(fx+eps)
px=exp(fx)
px=px/(1+px)
ly=rbinom(n=length(px),prob=px,size=1)
set.seed(1011)
cvob1=cv.glmnet(x,y)
animate_glmnet(cvob1)
}
\keyword{cv.glmnet}
\keyword{glmnet}
\keyword{lasso}
\keyword{machinelearning}
\keyword{seeAI}

